reduce duplications, fix minor bugs produced while doing manual rollback
