+++ frontend/connectoon/src/Containers/Recommendation/Recommendation.js
@@ -6,27 +6,16 @@ import { withRouter } from 'react-router';
 
 import * as actionCreators from '../../store/actions/index';
 import WorkList from '../../Components/WorkList/WorkList';
+import { stateConstructor, stateUpdator } from '../WorkPaginationTools/tools';
 
 import './Recommendation.css';
 
 class Recommendation extends Component {
   constructor(props) {
     super(props);
-    let subjectRows;
-    const worksInRow = 4;
-    const rowIncrement = 2;
-    const pageRowIncrement = 5;
-    const { location } = this.props;
-    if (location.state && location.state.subjectRows) {
-      subjectRows = location.state.subjectRows;
-    } else {
-      subjectRows = [1, 1];
-    }
-    const requestWorks = subjectRows.map((rows) => { return [0, worksInRow * (rows + pageRowIncrement)]; });
-    this.state = {
-      subjectRows, requestWorks, worksInRow, rowIncrement, pageRowIncrement,
-    };
-    const { onGetRecWorks } = this.props;
+    const { location, onGetRecWorks } = this.props;
+    this.state = stateConstructor(location);
+    const { requestWorks } = this.state;
     onGetRecWorks(requestWorks);
   }
 
@@ -35,20 +24,7 @@ class Recommendation extends Component {
   }
 
   onClickMore = (listId) => {
-    const {
-      subjectRows, requestWorks, worksInRow, rowIncrement, pageRowIncrement,
-    } = this.state;
-    subjectRows[listId] += rowIncrement;
-    const newRequestWorks = [];
-    let fetchMore = false;
-    requestWorks.forEach((requestWork, idx) => {
-      if (worksInRow * (subjectRows[idx] + rowIncrement) >= requestWork[1]) {
-        fetchMore = true;
-        newRequestWorks.push([requestWork[1], requestWork[1] + worksInRow * pageRowIncrement]);
-      } else {
-        newRequestWorks.push([requestWork[1], requestWork[1]]);
-      }
-    });
+    const { subjectRows, newRequestWorks, fetchMore } = stateUpdator(listId, this.state);
     this.setState({ subjectRows, requestWorks: newRequestWorks });
     if (fetchMore) {
       this.props.onGetMainWorks(newRequestWorks);
